changed from int to double

+++ src/TeamOneMiniProject/Calculations.java
@@ -2,8 +2,8 @@ package TeamOneMiniProject;
 
 public class Calculations {
 
-    public static int calculateSumHours (Member[] members){
-        int sum = 0;
+    public static Double calculateSumHours (Member[] members){
+        Double  sum = 0.0;
         for (int i = 0; i < members.length; i++){
             sum += members[i].getHoursWorked();
         }
